Some minor refactoring, mostly of tests
